requirements-common.txt
bowerstatic==0.9
dogpile.cache==0.5.7
dogpile.core==0.4.1
Chameleon==2.22
colander==1.0
deform==2.0a2
iso8601==0.1.10
Mako==1.0.1
MarkupSafe==0.23
meld3==1.0.2
Paste==2.0.2
PasteDeploy==1.5.2
peppercorn==0.5
Pygments==2.0.2
pyramid==1.5.7
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4
pyramid-layout==1.0
pyramid-mako==1.0.2
repoze.lru==0.6
rfc6266==0.0.4
six==1.9.0
translationstring==1.3
venusian==1.0
waitress==0.8.9
WebOb==1.4.1
wheel==0.24.0
zc.buildout==2.3.1
zope.deprecation==4.1.2
zope.exceptions==4.0.7
zope.interface==4.1.2
